backend/requirements.txt
# Core Web Framework
fastapi>=0.109.2
uvicorn[standard]>=0.27.1
python-multipart>=0.0.9
pydantic>=2.6.1
pydantic-settings>=2.0.0
python-dotenv>=1.0.1

# Data Processing (pre-compiled wheels for ARM64)
numpy>=1.26.3
pandas>=2.2.0
pyarrow>=15.0.0  # For parquet support
openpyxl>=3.1.2  # For Excel support

# Machine Learning
scikit-learn>=1.4.0
scipy>=1.12.0
joblib>=1.3.2
optuna>=3.5.0  # For hyperparameter optimization
lightgbm>=4.3.0  # For gradient boosting
xgboost>=2.0.3  # For gradient boosting
catboost>=1.2.2  # For gradient boosting

# AI/ML Services
openai>=1.12.0
httpx>=0.24.1  # Required by OpenAI client
anyio>=4.0.0   # Required by FastAPI
python-jose[cryptography]>=3.3.0  # For JWT handling
passlib[bcrypt]>=1.7.4  # For password hashing

# File Handling
aiofiles>=23.2.1  # For async file operations
python-magic>=0.4.27  # For file type detection

# Data Visualization
matplotlib>=3.8.2
seaborn>=0.13.2
plotly>=5.18.0

# Report Generation
Jinja2>=3.1.3
pdfkit>=1.0.0
python-docx>=1.1.0
markdown>=3.5.2

# API and Environment
kaggle>=1.6.6
requests>=2.31.0  # For URL downloads
GitPython>=3.1.44

# Development Tools
pytest>=8.0.1
black>=24.1.1
isort>=5.13.2
mypy>=1.8.0
pylint>=3.0.3

# SQLAlchemy and related dependencies
sqlalchemy>=2.0.0
alembic>=1.12.0

# Database drivers
psycopg2-binary>=2.9.9